refactor: extract HotkeyActionDispatcher.swift from HotkeySupportClasses.swift
Splits the `HotkeyActionDispatcher` @MainActor class (~39 lines) out
of HotkeySupportClasses.swift into its own file so that file holds
only the `HotkeySettingsBinder`.

Byte-identical move of lines 47-85. Both files import Combine +
Foundation.

HotkeySupportClasses.swift: 85 → 45 lines. Tests: 667.
